chore: move some helper to ha folder
