+++ src/ha/data/ws-templates.ts
